rationalise tests for diff api functions
